python-backend/requirements.txt
Flask
gunicorn
assemblyai
requests
flask-cors
python-docx
gTTS
nltk
